EDX-33 | SCH - View School Inbox Overall backend changes Test case fixes
