+++ backend/src/routes/user.js
@@ -2,7 +2,7 @@
 
 const passport = require('passport');
 const express = require('express');
-const { getUserInfo } = require('../components/user');
+const {getUserInfo} = require('../components/user');
 
 const router = express.Router();
 const auth = require('../components/auth');
